chore: move profiles util mapper to util dir
